% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Solve Knapsack Problems by Using Brute Force}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{data.frame with two variables (item weight and item value) that is to be used for the brute force search.}

\item{W}{limit for the amount of weight that the knapsack can take on.}
}
\value{
list containing the maximum knapsack value and the positions of the associated items.
}
\description{
A function that uses the brute force approach for finding the maximum value for a knapsack problem.
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w = sample(1:4000, size = n, replace = TRUE),
                             v <- runif(n = n, 0, 10000))

brute_force_knapsack(x = knapsack_objects[1: 8,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 2000)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 2000)
}
