% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Solve Knapsack Problems by Using a Greedy Heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data.frame with two variables (item weight and item value).}

\item{W}{limit for the amount of weight that the knapsack can take on.}
}
\value{
list containing the estimated maximum knapsack value and the positions of the associated items.
}
\description{
A function that uses a greedy heuristic for estimating the maximum value for a knapsack problem.
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w = sample(1:4000, size = n, replace = TRUE),
                             v <- runif(n = n, 0, 10000))

greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
greedy_knapsack(x = knapsack_objects[1:1200,], W = 2000)
}
