% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\docType{package}
\name{knapsack}
\alias{knapsack}
\title{Knapsack: A package for solving knapsack problems}
\description{
The knapsack package provides two functions for solving knapsack problems:
brute_force_knapsack(), and greedy_knapsack().
}
\section{brute_force_knapsack()}{

The brute_force_knapsack function uses the brute force approach for finding the maximum value for a knapsack problem.
For more details, write ?brute_force_knapsack.
}

\section{greedy_knapsack()}{

The greedy_knapsack function uses a greedy heuristic for estimating the maximum value for a knapsack problem..
For more details, write ?greedy_knapsack.
}

