% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Weights and values for items in a knapsack (2000 items).}
\format{

}
\usage{
knapsack_objects
}
\description{
A dataset containing the weights and values for items in a knapsack with size 2000
}
\keyword{datasets}
