% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects2.R
\docType{data}
\name{knapsack_objects2}
\alias{knapsack_objects2}
\title{Weights and values for items in a knapsack (1000000 items).}
\format{

}
\usage{
knapsack_objects2
}
\description{
A dataset containing the weights and values for items in a knapsack with size 1000000
}
\keyword{datasets}
